Moved bottom_up, use, and walk to core
