+++ sympy/__init__.py
@@ -60,7 +60,7 @@ from .core import (sympify, SympifyError, cacheit, Basic, Atom,
         expand_func, expand_trig, expand_complex, expand_multinomial, nfloat,
         expand_power_base, expand_power_exp, arity, PrecisionExhausted, N,
         evalf, Tuple, Dict, gcd_terms, factor_terms, factor_nc, evaluate,
-        Catalan, EulerGamma, GoldenRatio, TribonacciConstant)
+        Catalan, EulerGamma, GoldenRatio, TribonacciConstant, bottom_up, use)
 
 from .logic import (to_cnf, to_dnf, to_nnf, And, Or, Not, Xor, Nand, Nor,
         Implies, Equivalent, ITE, POSform, SOPform, simplify_logic, bool_map,
@@ -153,8 +153,8 @@ from .discrete import (fft, ifft, ntt, intt, fwht, ifwht, mobius_transform,
         intersecting_product)
 
 from .simplify import (simplify, hypersimp, hypersimilar, logcombine,
-        separatevars, posify, besselsimp, kroneckersimp, signsimp, bottom_up,
-        nsimplify, FU, fu, sqrtdenest, cse, use, epath, EPath, hyperexpand,
+        separatevars, posify, besselsimp, kroneckersimp, signsimp,
+        nsimplify, FU, fu, sqrtdenest, cse, epath, EPath, hyperexpand,
         collect, rcollect, radsimp, collect_const, fraction, numer, denom,
         trigsimp, exptrigsimp, powsimp, powdenest, combsimp, gammasimp,
         ratsimp, ratsimpmodprime)
@@ -269,7 +269,7 @@ __all__ = [
     'expand_complex', 'expand_multinomial', 'nfloat', 'expand_power_base',
     'expand_power_exp', 'arity', 'PrecisionExhausted', 'N', 'evalf', 'Tuple',
     'Dict', 'gcd_terms', 'factor_terms', 'factor_nc', 'evaluate', 'Catalan',
-    'EulerGamma', 'GoldenRatio', 'TribonacciConstant',
+    'EulerGamma', 'GoldenRatio', 'TribonacciConstant', 'bottom_up', 'use',
 
     # sympy.logic
     'to_cnf', 'to_dnf', 'to_nnf', 'And', 'Or', 'Not', 'Xor', 'Nand', 'Nor',
@@ -379,8 +379,8 @@ __all__ = [
 
     # sympy.simplify
     'simplify', 'hypersimp', 'hypersimilar', 'logcombine', 'separatevars',
-    'posify', 'besselsimp', 'kroneckersimp', 'signsimp', 'bottom_up',
-    'nsimplify', 'FU', 'fu', 'sqrtdenest', 'cse', 'use', 'epath', 'EPath',
+    'posify', 'besselsimp', 'kroneckersimp', 'signsimp',
+    'nsimplify', 'FU', 'fu', 'sqrtdenest', 'cse', 'epath', 'EPath',
     'hyperexpand', 'collect', 'rcollect', 'radsimp', 'collect_const',
     'fraction', 'numer', 'denom', 'trigsimp', 'exptrigsimp', 'powsimp',
     'powdenest', 'combsimp', 'gammasimp', 'ratsimp', 'ratsimpmodprime',
