+++ sympy/functions/elementary/miscellaneous.py
@@ -1,6 +1,7 @@
 from sympy.core import Function, S, sympify
 from sympy.utilities.iterables import sift
 from sympy.core.add import Add
+from sympy.core.basic import walk
 from sympy.core.containers import Tuple
 from sympy.core.compatibility import ordered
 from sympy.core.operations import LatticeOp, ShortCircuit
@@ -17,6 +18,7 @@ from sympy.core.symbol import Dummy
 from sympy.core.rules import Transform
 from sympy.core.logic import fuzzy_and, fuzzy_or, _torf
 from sympy.logic.boolalg import And, Or
+
 
 def _minmax_as_Piecewise(op, *args):
     # helper for Min/Max rewrite as Piecewise
@@ -434,8 +436,6 @@ class MinMaxBase(Expr, LatticeOp):
         >>> Min(a, Max(b, Min(c, d, Max(a, e))))
         Min(a, Max(b, Min(a, c, d)))
         """
-        from sympy.simplify.simplify import walk
-
         if not args:
             return args
         args = list(ordered(args))
