+++ sympy/polys/polytools.py
@@ -7,7 +7,7 @@ from operator import mul
 from sympy.core import (
     S, Expr, I, Integer, Add, Tuple
 )
-from sympy.core.basic import Basic, preorder_traversal
+from sympy.core.basic import Basic, preorder_traversal, bottom_up
 from sympy.core.compatibility import iterable, ordered
 from sympy.core.decorators import _sympifyit
 from sympy.core.evalf import pure_complex
@@ -48,6 +48,7 @@ from sympy.polys.rationaltools import together
 from sympy.polys.rootisolation import dup_isolate_real_roots_list
 from sympy.utilities import group, sift, public, filldedent
 from sympy.utilities.exceptions import SymPyDeprecationWarning
+
 
 # Required to avoid errors
 import sympy.polys
@@ -6365,7 +6366,6 @@ def factor(f, *gens, deep=False, **args):
     """
     f = sympify(f)
     if deep:
-        from sympy.simplify.simplify import bottom_up
         def _try_factor(expr):
             """
             Factor, but avoid changing the expression when unable to.
